accept "numeric"  shade tolerances

+++ Biomass_core.R
@@ -454,7 +454,7 @@ Init <- function(sim, verbose = getOption("LandR.verbose", TRUE)) {
   species <- setDT(sim$species)[, speciesCode := as.factor(species)]
   LandR::assertColumns(species,
                        c(species = "character", Area = "factor", longevity = "integer",
-                         sexualmature = "integer", shadetolerance = "integer",
+                         sexualmature = "integer", shadetolerance = "numeric",
                          firetolerance = "integer", seeddistance_eff = "integer",
                          seeddistance_max = "integer", resproutprob = "numeric",
                          resproutage_min = "integer", resproutage_max = "integer",
